requirements.txt
requests
pillow
selenium
beautifulsoup4